Rename Insertion.cpp to InsertionTraversalDeletion.cpp
Explanation:

Node Class: The Node class is used to define the structure of each node in the linked list. It contains a data field (int data) to store the actual value of the node and a Node* next pointer to point to the next node in the sequence.

Creating Nodes: In the main function, we create three nodes (node1, node2, and node3) with values 10, 20, and 30, respectively.

Linking Nodes: We link the nodes together by setting the next pointer of each node to point to the next node in the sequence.

Traversing the Linked List: We use a while loop to traverse the linked list starting from the head (node1) and print the data of each node.